+++ core/src/builtin_widgets/smooth_layout.rs
@@ -612,7 +612,7 @@ mod tests {
 
     assert_widget_eq_image!(
       WidgetTester::new(self::column! {
-        clamp: BoxClamp::fixed_width(100.),
+        clamp: BoxClamp::default().with_max_width(100.),
         align_items: Align::Center,
         // If no initial value is provided, the widget should start at its real place.
         @SmoothX {
@@ -631,6 +631,7 @@ mod tests {
           init_value: HAnchor::Right(0f32.into()),
           @red_block_10_x_10()
         }
+        @SizedBox { size: Size::new(100., 10.) }
       })
       .with_wnd_size(Size::new(100., 30.))
       .on_initd(|wnd| wnd.set_flags(WindowFlags::ANIMATIONS)),
@@ -645,7 +646,7 @@ mod tests {
 
     assert_widget_eq_image!(
       WidgetTester::new(self::row! {
-        clamp: BoxClamp::fixed_height(100.),
+        clamp: BoxClamp::default().with_max_height(100.),
         align_items: Align::Center,
         // If no initial value is provided, the widget should start at its real place.
         @SmoothY {
@@ -664,6 +665,7 @@ mod tests {
           init_value: VAnchor::Bottom(0f32.into()),
           @red_block_10_x_10()
         }
+        @SizedBox { size: Size::new(10., 100.) }
       })
       .with_wnd_size(Size::new(30., 100.))
       .on_initd(|wnd| wnd.set_flags(WindowFlags::ANIMATIONS)),